collegeERPsystem/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
drf-spectacular>=0.27
psycopg2-binary
django-environ
django-filter
celery>=5.2
redis
django-cors-headers
Pillow
python-dotenv
pytest
pytest-django
factory-boy
gunicorn
